Move RegionManagerExtensions.Add extension method for IRegionCollection to IRegionCollection.
