+++ Source/Wpf/Prism.Wpf.Tests/Prism.Wpf.Tests.csproj
@@ -248,7 +248,6 @@
     <Compile Include="Regions\RegionFixture.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Regions\RegionManagerExtensionsFixture.cs" />
     <Compile Include="Regions\RegionManagerFixture.cs">
       <SubType>Code</SubType>
     </Compile>
