+++ Source/Wpf/Prism.Wpf/Regions/RegionManager.cs
@@ -392,6 +392,30 @@ namespace Prism.Regions
                 UpdateRegions();
 
                 return GetRegionByName(regionName) != null;
+            } 
+            
+            /// <summary>
+            /// Adds a region to the regionmanager with the name received as argument.
+            /// </summary>
+            /// <param name="regionName">The name to be given to the region.</param>
+            /// <param name="region">The region to be added to the regionmanager.</param>        
+            /// <exception cref="ArgumentNullException">Thrown if <paramref name="region"/> is <see langword="null"/>.</exception>
+            /// <exception cref="ArgumentException">Thrown if <paramref name="regionName"/> and <paramref name="region"/>'s name do not match and the <paramref name="region"/> <see cref="IRegion.Name"/> is not <see langword="null"/>.</exception>
+            public void Add(string regionName, IRegion region)
+            {
+                if (region == null) throw new ArgumentNullException("region");
+
+                if (region.Name != null && region.Name != regionName)
+                {
+                    throw new ArgumentException(string.Format(CultureInfo.CurrentCulture, Resources.RegionManagerWithDifferentNameException, region.Name, regionName), "regionName");
+                }
+
+                if (region.Name == null)
+                {
+                    region.Name = regionName;
+                }
+
+                Add(region);
             }
 
             private IRegion GetRegionByName(string regionName)
