+++ Source/Wpf/Prism.Wpf/Regions/RegionManagerExtensions.cs
@@ -13,32 +13,6 @@ namespace Prism.Regions
     public static class RegionManagerExtensions
     {
         /// <summary>
-        /// Adds a region to the regionmanager with the name received as argument.
-        /// </summary>
-        /// <param name="regionCollection">The regionmanager's collection of regions.</param>
-        /// <param name="regionName">The name to be given to the region.</param>
-        /// <param name="region">The region to be added to the regionmanager.</param>        
-        /// <exception cref="ArgumentNullException">Thrown if <paramref name="region"/> or <paramref name="regionCollection"/> is <see langword="null"/>.</exception>
-        /// <exception cref="ArgumentException">Thrown if <paramref name="regionName"/> and <paramref name="region"/>'s name do not match and the <paramref name="region"/> <see cref="IRegion.Name"/> is not <see langword="null"/>.</exception>
-        public static void Add(this IRegionCollection regionCollection, string regionName, IRegion region)
-        {
-            if (region == null) throw new ArgumentNullException("region");
-            if (regionCollection == null) throw new ArgumentNullException("regionCollection");
-
-            if (region.Name != null && region.Name != regionName)
-            {
-                throw new ArgumentException(string.Format(CultureInfo.CurrentCulture, Resources.RegionManagerWithDifferentNameException, region.Name, regionName), "regionName");
-            }
-
-            if (region.Name == null)
-            {
-                region.Name = regionName;
-            }
-
-            regionCollection.Add(region);
-        }
-
-        /// <summary>
         /// Navigates the specified region manager.
         /// </summary>
         /// <param name="regionManager">The regionmanager that this extension method effects.</param>
